gae: breaking change: ordered /timeseries output
I did not want to break this,
but given the access numbers (<50 requests
in the last 24 hours) I think it is fair
to  better break that early than later.

The output of individual samples should be
_ordered_, i.e. in an array, so that once
can, for instance, easily access the last
sample.

Other consumer need to parse through the
individual timestamps anyway, it is
unlikely that consumers index a
map with an _exact_ time string.

This resolves issue #12.