epsilon merge tool: lost column support
when between base and extension a column is lost
(when it appears in base, but not in extension)
then forward-fill that column into the extension
data set, using the last value from the base.

this is to accomodate the loss of the column
for AGS 16056 in the RKI data set around
October 2021 (unclear what happened exactly).