build-rki-csvs: remove strict ags set check
In oct 2021 the AGS 16056 disappeared from
the RKI data set. That is, lacking_wrt_ref

was not

  set([11000, 3152])

anymore but

  set([11000, 3152, 16056])

which is when the program expectedly
crashed, leading up to issue #1748.

Now, deal with unexpected loss of a
column. This is accomodated for in the
epsilon merge tool which forward-fills
a column when it's in the base but not in
the extension.